+++ client/src/locales/en.json
@@ -8,7 +8,8 @@
 		"labels": {
 			"responseTime": "Response time",
 			"active": "Active",
-			"resolved": "Resolved"
+			"resolved": "Resolved",
+			"na": "N/A"
 		},
 		"buttons": {
 			"toggleTheme": "Toggles light & dark",
@@ -406,7 +407,8 @@
 				},
 				"status": {
 					"active": "Active",
-					"resolved": "Resolved"
+					"resolved": "Resolved",
+			"na": "N/A"
 				},
 				"actions": {
 					"details": "Details",
@@ -452,20 +454,6 @@
 			}
 		}
 	},
-	"incidentsTableNoIncidents": "No incidents recorded",
-	"incidentsTablePaginationLabel": "incidents",
-	"incidentsTableMonitorName": "Monitor Name",
-	"incidentsTableStatus": "Status",
-	"incidentsTableStatusCode": "Status Code",
-	"incidentsTableMessage": "Message",
-	"incidentsOptionsHeader": "Incidents for:",
-	"incidentsOptionsHeaderFilterBy": "Filter by:",
-	"incidentsOptionsHeaderFilterAll": "All",
-	"incidentsOptionsHeaderShow": "Show:",
-	"incidentsOptionsHeaderLastHour": "Last hour",
-	"incidentsOptionsHeaderLastDay": "Last day",
-	"incidentsOptionsHeaderLastWeek": "Last week",
-	"incidentsOptionsPlaceholderAllServers": "All servers",
 	"used": "Used",
 	"total": "Total",
 	"status": "Status",
@@ -648,7 +636,6 @@
 	"statusBreadCrumbsDetails": "Details",
 	"commonSaving": "Saving...",
 	"navControls": "Controls",
-	"incidentsPageTitle": "Incidents",
 	"passwordPanel": {
 		"currentPassword": "Current password",
 		"newPassword": "New password",
@@ -1002,61 +989,9 @@
 	},
 	"chooseGame": "Choose game",
 	"incidentsPage": {
-		"resolved": "Resolved",
-		"active": "Active",
-		"startTime": "Start Time",
-		"endTime": "End Time",
-		"resolutionType": "Resolution Type",
-		"incidentsActivePanelError": "Error loading active incidents",
-		"allSystemsAreOperational": "All Systems Are Operational",
-		"incidentsActivePanelTitle": "Total Active Incidents",
-		"incidentsLatestPanelError": "Error loading latest incidents",
-		"incidentsLatestPanelTitle": "Latest Incidents",
-		"incidentsLatestPanelEmpty": "No recent incidents",
-		"incidentItemMonitor": "Monitor",
-		"unknownMonitor": "Unknown Monitor",
-		"incidentsStatisticsPanelError": "Error loading statistics",
-		"incidentsStatisticsPanelTitle": "General Statistics",
-		"totalIncidents": "Total Incidents",
-		"mostAffectedMonitor": "Most Affected Monitor",
-		"avgResolutionTime": "Avg. Resolution Time",
-		"incidentResolvedSuccessfully": "Incident resolved successfully",
-		"errorFetchingIncidents": "Error fetching incidents",
-		"errorFetchingIncident": "Error fetching incident",
-		"noIncidentIdProvided": "No incident ID provided",
-		"errorFetchingIncidentSummary": "Error fetching incident summary",
-		"incidentsOptionsHeaderFilterManual": "Manual",
-		"incidentsOptionsHeaderFilterAutomatic": "Automatic",
-		"incidentsOptionsHeaderFilterAll": "All",
-		"incidentsOptionsHeaderFilterActive": "Active",
-		"incidentsOptionsHeaderFilterResolved": "Resolved",
-		"incidentsTableActionResolveManually": "Resolve Manually",
-		"incidentsTableActionDetails": "Details",
-		"incidentsTableActionGoToMonitor": "Go to monitor",
 		"resolveIncidentDialogTitle": "Resolve Incident",
-		"resolveIncidentDialogDescription": "Are you sure you want to resolve this incident manually?",
 		"resolveIncidentDialogCommentLabel": "Comment (optional)",
 		"resolveIncidentDialogCommentPlaceholder": "Add a comment about the resolution...",
-		"resolveIncidentDialogConfirm": "Resolve",
-		"hours": "hours",
-		"incidentNotFound": "Incident not found",
-		"incidentDetails": "INCIDENT DETAILS",
-		"overview": "OVERVIEW",
-		"URL": "URL",
-		"incidentAnalysis": "INCIDENT ANALYSIS",
-		"timeline": "Timeline",
-		"startedAt": "Started at",
-		"endedAt": "Ended at",
-		"downtime": "Downtime",
-		"errorDetails": "Error details",
-		"resolutionDetails": "Resolution details",
-		"statusCode": "Status code",
-		"resolutionMethod": "Resolution method",
-		"resolvedBy": "Resolved by",
-		"comment": "Comment",
-		"noCommentProvided": "No comment provided",
-		"loadingIncidentDetails": "Loading incident details...",
-		"none": "None",
-		"message": "Message"
+		"resolveIncidentDialogConfirm": "Resolve"
 	}
 }
